+++ server/service/client_setup.go
@@ -1,61 +1,20 @@
-package client
+package service
 
 import (
-	"bytes"
-	"crypto/tls"
 	"encoding/json"
 	"fmt"
 	"io/ioutil"
 	"net/http"
 
 	"github.com/kolide/fleet/server/kolide"
-	"github.com/kolide/fleet/server/service"
 	"github.com/pkg/errors"
 )
-
-type Client struct {
-	addr string
-	http *http.Client
-}
-
-func New(addr string, insecureSkipVerify bool) (*Client, error) {
-	httpClient := &http.Client{
-		Transport: &http.Transport{
-			TLSClientConfig: &tls.Config{InsecureSkipVerify: insecureSkipVerify},
-		},
-	}
-
-	return &Client{
-		addr: addr,
-		http: httpClient,
-	}, nil
-}
-
-func (c *Client) Do(verb, path string, params interface{}) (*http.Response, error) {
-	b, err := json.Marshal(params)
-	if err != nil {
-		return nil, errors.Wrap(err, "error marshaling json")
-	}
-
-	request, err := http.NewRequest(
-		verb,
-		c.addr+path,
-		bytes.NewBuffer(b),
-	)
-	if err != nil {
-		return nil, errors.Wrap(err, "error creating request object")
-	}
-	request.Header.Set("content-type", "application/json")
-	request.Header.Set("accept", "application/json")
-
-	return c.http.Do(request)
-}
 
 // Setup attempts to setup the current Fleet instance. If setup is successful,
 // an auth token is returned.
 func (c *Client) Setup(email, password, org string) (string, error) {
 	t := true
-	params := service.SetupRequest{
+	params := setupRequest{
 		Admin: &kolide.UserPayload{
 			Admin:    &t,
 			Username: &email,
@@ -89,7 +48,7 @@ func (c *Client) Setup(email, password, org string) (string, error) {
 		return "", errors.Wrap(err, "error reading response body")
 	}
 
-	var responseBody service.SetupResponse
+	var responseBody setupResponse
 	err = json.Unmarshal(responeBytes, &responseBody)
 	if err != nil {
 		return "", errors.Wrap(err, "error decoding HTTP response body")
