+++ server/service/transport_setup.go
@@ -7,7 +7,7 @@ import (
 )
 
 func decodeSetupRequest(ctx context.Context, r *http.Request) (interface{}, error) {
-	var req SetupRequest
+	var req setupRequest
 	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
 		return nil, err
 	}
